Project-frontend/PF/requirements.txt
asgiref==3.5.2
Django==4.0.8
django-cors-headers==3.13.0
django-dirtyfields==1.9.0
django-filter==22.1
django-location-field==2.1.0
django-password-validators==1.7.0
djangorestframework==3.14.0
djangorestframework-filters==1.0.0.dev0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
install==1.3.5
Pillow==9.3.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
python-dateutil==2.8.2
regex==2022.10.31
